+++ src/shared/query/useChatList.ts
@@ -1,12 +1,8 @@
 import type { AxiosError, AxiosResponse } from 'axios';
 import { useMemo } from 'react';
-import { useMutation, useQuery, useQueryClient } from 'react-query';
-import { createRoom, getChatList, searchChatList } from 'shared/api/chatApi';
+import { useQuery } from 'react-query';
+import { getChatList, searchChatList } from 'shared/api/chatApi';
 import queryKeys from 'shared/constants/queryKeys';
-
-type Req = {
-  userList: number[];
-};
 
 type Res = {
   roomId: number;
@@ -17,21 +13,14 @@ type Res = {
 }[];
 
 export default function useChatList(name?: string) {
-  const queryClient = useQueryClient();
   const { data, refetch } = useQuery<AxiosResponse<Res>, AxiosError>(
-    [queryKeys.CHATS, name],
+    [queryKeys.CHATS],
     name.length > 0 ? () => searchChatList(name) : getChatList,
     {
-      enabled: !!name,
+      refetchOnWindowFocus: false,
     }
   );
   const chatList = useMemo(() => data?.data, [data]);
 
-  const { mutate } = useMutation<AxiosResponse, AxiosError, Req>(
-    ({ userList }) => createRoom(userList),
-    {
-      onSuccess: res => queryClient.invalidateQueries([queryKeys.CHATS]),
-    }
-  );
-  return { chatList, refetch, mutate };
+  return { chatList, refetch };
 }
